project update
+ add placesServiceMock
* move some files for general betterness
