+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon-40@2x.png
