+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon-60@2x.png
