+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon-60@3x.png
