+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon-72@2x.png
