+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon-76@2x.png
