+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon-small-50@2x.png
