+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon-small@2x.png
