+++ weg-li/App/Resources/Assets.xcassets/AppIcon.appiconset/icon@2x.png
