+++ weg-li/App/Resources/Assets.xcassets/Contents.json
