+++ weg-li/UI/MailContentView.swift
